% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fill_query}
\alias{fill_query}
\title{Get a names list of ontology and id by id}
\usage{
fill_query(all, query)
}
\arguments{
\item{all}{all ontology id to plot tree}

\item{query}{query ontology id to fill}
}
\value{
a color object to fill query in onto_plot
}
\description{
Get a names list of ontology and id by id
}
\examples{
\dontrun{
fill_query(all = c("CL0000548", "CL:0000066", "CL:0000082"), query = c("CL:0000082"))
}
}
