% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getOntologyName}
\alias{getOntologyName}
\title{Get a names list of ontology and id by id}
\usage{
getOntologyName(ont, onto_id)
}
\arguments{
\item{ont}{ontology object}

\item{onto_id}{character vector of ontology ids}
}
\value{
a named list mapping ontology id and ontology name
}
\description{
Get a names list of ontology and id by id
}
\examples{
\dontrun{
getOntologyName(onto_id = "epithelial cell of lung", ont = ont)
}
}
