% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOntoMapping}
\alias{getOntoMapping}
\alias{get.onto.mapping}
\title{match ontology terms by direct mapping and mapping descendants to ancestors}
\usage{
get.onto.mapping(ont, onts1, onts2)
}
\arguments{
\item{ont}{the ontology object from get_OBO}

\item{onts1}{a named list looks like ontology_id:ontology name}

\item{onts2}{a named list looks like ontology_id:ontology name}
}
\value{
a names list for ontology id mapping
}
\description{
match ontology terms by direct mapping and mapping descendants to ancestors
}
