% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOntoMultiMapping}
\alias{getOntoMultiMapping}
\title{Match descendants to ancestors in multiple ontology id lists}
\usage{
getOntoMultiMapping(ont, onts)
}
\arguments{
\item{ont}{ontologyIndex object}

\item{onts}{named list of ontology ids}
}
\value{
a named character of mapping from:mapping to
}
\description{
Match descendants to ancestors in multiple ontology id lists
}
\examples{
\dontrun{
getOntoMultiMapping(ont = ont, onts = c("CL0000548", "CL:0000066", "CL:0000082"))
}
}
