% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getSeuratRds}
\alias{getSeuratRds}
\title{read in anndata files as a named list of anndata object}
\usage{
getSeuratRds(metadata, sep)
}
\arguments{
\item{metadata}{a metadata file indicating name, path to seurat rds file}

\item{sep}{sep of the metadata file}
}
\description{
read in anndata files as a named list of anndata object
}
