% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ontoMatch}
\alias{ontoMatch}
\title{Core function of scOntoMatch
Match the ontology annotation of several obj files}
\usage{
ontoMatch(adata1, adata2, anno_col, onto_id_col, ont)
}
\arguments{
\item{adata1}{one anndata object}

\item{adata2}{the other anndata object}

\item{anno_col}{the cell ontology text annotation column name}

\item{onto_id_col}{if also have ontology id column for direct mapping}

\item{ont}{ontology file loaded via get_OBO}
}
\value{
a list of obj files with annotation ontology mapped to each-other in obs[['cell_type_mapped_ontology']]
}
\description{
Core function of scOntoMatch
Match the ontology annotation of several obj files
}
