% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ontoMinimal}
\alias{ontoMinimal}
\title{get the minimal ontology tree of a dataset by reducing descendant terms to ancestor terms
return  obj meta.data[["cell_ontology_base"]] storing the reduced ontology annotation}
\usage{
ontoMinimal(obj, ont, anno_col, onto_id_col)
}
\arguments{
\item{obj}{the seurat object}

\item{ont}{ontologyIndex object}

\item{anno_col}{the cell ontology text annotation column name}

\item{onto_id_col}{if also have ontology id column for direct mapping}
}
\value{
an anndata object with .obs[["cell_ontology_base"]]
}
\description{
get the minimal ontology tree of a dataset by reducing descendant terms to ancestor terms
return  obj meta.data[["cell_ontology_base"]] storing the reduced ontology annotation
}
