% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ontoMultiMatch}
\alias{ontoMultiMatch}
\title{Core function of scOntoMatch
Match the ontology annotation of multiple seurat objects}
\usage{
ontoMultiMatch(obj_list, anno_col, onto_id_col, ont)
}
\arguments{
\item{obj_list}{a namesd list of seurat objects to match}

\item{anno_col}{the cell ontology text annotation column name}

\item{onto_id_col}{if also have ontology id column for direct mapping}

\item{ont}{ontologyIndex object}
}
\value{
a list of seurat objects with annotation ontology mapped to each-other in obs[['cell_ontology_mapped']]
}
\description{
Core function of scOntoMatch
Match the ontology annotation of multiple seurat objects
}
\examples{
\dontrun{
ontoMultiMatch(seurat_obj_list, ont, "ontology_name", 'ontology_id')
}
}
