% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ontoTranslate}
\alias{ontoTranslate}
\title{translate named list of obj_list to named list of cell ontology ids per obj}
\usage{
ontoTranslate(obj_list, ont, onto_id_col, anno_col)
}
\arguments{
\item{obj_list}{a named list of seurat object}

\item{ont}{ontologyIndex object}

\item{onto_id_col}{if also have ontology id column for direct mapping}

\item{anno_col}{the cell ontology text annotation column name}
}
\value{
a named list of cell ontology ids
}
\description{
translate named list of obj_list to named list of cell ontology ids per obj
}
\examples{
\dontrun{
ontoTranslate(seurat_obj_list, ont, 'ontology_name', 'ontology_id')
}
}
