% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotOntoTree}
\alias{plotOntoTree}
\alias{plot.onto.tree}
\title{Plot a tree representation of ontology terms}
\usage{
\method{plot}{onto.tree}(ont, onts, plot_ancestors = TRUE, ont_query, ...)
}
\arguments{
\item{ont}{ontology object}

\item{onts}{ontology ids to plot}

\item{plot_ancestors}{if plot ancestors or not}

\item{ont_query}{query ontology to highlight in the tree}
}
\value{
an ontology tree plot
}
\description{
Plot a tree representation of ontology terms
}
