% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOntologyId}
\alias{getOntologyId}
\title{Get a names list of ontology and id by name}
\usage{
getOntologyId(
  adata,
  anno_col = "authors_cell_type_-_ontology_labels",
  obo_object
)
}
\arguments{
\item{adata}{adata object}

\item{anno_col}{the cell ontology text annotation column name}

\item{obo_object}{obo object}
}
\value{
a named list mapping ontology id and ontology name
}
\description{
Get a names list of ontology and id by name
}
