% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getOntologyId}
\alias{getOntologyId}
\title{Get a names list of ontology and id by name}
\usage{
getOntologyId(ont, onto_name)
}
\arguments{
\item{ont}{ontology object}

\item{onto_name}{character vector of ontology names}
}
\value{
a named list mapping ontology id and ontology name
}
\description{
Get a names list of ontology and id by name
}
\examples{
\dontrun{
getOntologyId(onto_name = "epithelial cell of lung", ont = ont)
}
}
