% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOntologyName}
\alias{getOntologyName}
\title{Get a names list of ontology and id by id}
\usage{
getOntologyName(
  adata,
  onto_id_col = "authors_cell_type_-_ontology_labels_ontology",
  ont
)
}
\arguments{
\item{adata}{adata object}

\item{onto_id_col}{if also have ontology id column for direct mapping}

\item{ont}{ontology object}
}
\value{
a named list mapping ontology id and ontology name
}
\description{
Get a names list of ontology and id by id
}
