% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOntologyName}
\alias{getOntologyName}
\title{Get a names list of ontology and id by id}
\usage{
getOntologyName(onto_id, ont)
}
\arguments{
\item{onto_id}{character vector of ontology ids}

\item{ont}{ontology object}
}
\value{
a named list mapping ontology id and ontology name
}
\description{
Get a names list of ontology and id by id
}
