% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getOntoMapping}
\alias{getOntoMapping}
\title{match ontology terms by direct mapping and mapping descendants to ancestors}
\usage{
getOntoMapping(ont, onts1, onts2)
}
\arguments{
\item{ont}{the ontology object from get_OBO}

\item{onts1}{a character vector of ontology id}

\item{onts2}{a character vector of ontology id}
}
\value{
a named list for ontology id mapping looks like ontology_id:ontology_id
}
\description{
match ontology terms by direct mapping and mapping descendants to ancestors
}
\examples{
\dontrun{
getOntoMapping(ont = ont, onts1 = "CL:0000548", onts2 = c("CL0000548", "CL:0000066"))
}
}
