% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getSeuratRds}
\alias{getSeuratRds}
\title{read in seurat object .rds files as a named list of seurat object}
\usage{
getSeuratRds(metadata, sep)
}
\arguments{
\item{metadata}{a metadata file indicating name, path to 'seurat' rds file}

\item{sep}{sep of the metadata file}
}
\value{
a named list contains data name and the corresponding 'seurat' object
}
\description{
read in seurat object .rds files as a named list of seurat object
}
\examples{
\dontrun{
getSeuratRds(metadata = 'metadata.tsv', sep = '\t')
}
}
