% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ontoMatch}
\alias{ontoMatch}
\alias{onto.match}
\title{Core function of scOntoMatch
Match the ontology annotation of several adata files}
\usage{
onto.match(
  adata1,
  adata2,
  anno_col,
  onto_id_col,
  obo_file,
  propagate_relationships = c("is_a", "part_of")
)
}
\arguments{
\item{adata1}{one anndata file path}

\item{adata2}{the other anndata file path}

\item{anno_col}{the cell ontology text annotation column name}

\item{onto_if_col}{if also have ontology id column for direct mapping}
}
\value{
a list of adata files with annotation ontology mapped to each-other
}
\description{
Core function of scOntoMatch
Match the ontology annotation of several adata files
}
